Implement main application structure with notification service and profile display
