+++ lib/main_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'profile_page.dart';
+
+class MainScreen extends StatelessWidget {
+  const MainScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: ProfilePage(),
+    );
+  }
+}
